third_party/WebKit/LayoutTests/inspector/console/console-log-syntax-error-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token )
Tests that syntax errors are logged into console and doesn't cause browser crash.

error: TypeError: InspectorTest.dumpConsoleMessagesIgnoreStackFrames is not a function
